requirements.txt
yfinance==0.2.66
pandas==2.3.2
numpy==2.3.3
scipy==1.17.0
openpyxl==3.1.5
pyyaml==6.0.3
pyarrow==23.0.0
pydantic==2.12.3
pytest>=8.0.0
pandas-datareader>=0.10.0  # factor_exposure.py only
statsmodels>=0.14.0        # factor_exposure.py only
